+++ packages/companion-module-base/src/module-api/index.ts
@@ -6,7 +6,7 @@ export * from './enums.js'
 export * from './feedback.js'
 export * from './http.js'
 export * from './input.js'
-export * from './preset.js'
+export * from './preset/main.js'
 export {
 	SharedUdpSocket,
 	SharedUdpSocketEvents,
