+++ packages/companion-module-base/src/module-api/preset/main.ts
@@ -1,0 +1,12 @@
+import type { InstanceTypes } from '../base.js'
+import type { CompanionPresetDefinition } from './definition.js'
+
+export * from './definition.js'
+export * from './structure.js'
+
+/**
+ * The definitions of a group of feedbacks
+ */
+export type CompanionPresetDefinitions<TManifest extends InstanceTypes = InstanceTypes> = {
+	[id: string]: CompanionPresetDefinition<TManifest> | undefined
+}
